removed module folder from both ft-standard and storage-standard and updated examples
