+++ examples/fungible-token/test-contract-defi/src/lib.rs
@@ -1,7 +1,7 @@
 /*!
 Some hypothetical DeFi contract that will do smart things with the transferred tokens
 */
-use near_contract_standards::fungible_token::{receiver::FungibleTokenReceiver, Balance};
+use near_ft_standard::fungible_token::{receiver::FungibleTokenReceiver, Balance};
 use near_sdk::borsh::{BorshDeserialize, BorshSerialize};
 use near_sdk::json_types::U128;
 use near_sdk::{env, log, near_bindgen, require, AccountId, Gas, PanicOnDefault, PromiseOrValue};
