+++ near-contract-standards/src/event.rs
@@ -9,7 +9,6 @@ use near_sdk::serde_json;
 #[serde(rename_all = "snake_case")]
 pub(crate) enum NearEvent<'a> {
     Nep171(crate::non_fungible_token::events::Nep171Event<'a>),
-    //Nep141(near_fungible_token::fungible_token::events::Nep141Event<'a>),
 }
 
 impl<'a> NearEvent<'a> {
