+++ near-ft-standard/src/core.rs
