+++ near-ft-standard/src/events.rs
