+++ near-ft-standard/src/lib.rs
@@ -1,2 +1,18 @@
-pub mod fungible_token;
+//! This module represents a Fungible Token standard.
+//!
+//! # Examples
+//! See [`FungibleTokenCore`] and [`FungibleTokenResolver`] for example usage and [`FungibleToken`]
+//! for core standard implementation.
+pub mod core;
+pub mod core_impl;
 pub(crate) mod event;
+pub mod events;
+pub mod macros;
+pub mod metadata;
+pub mod receiver;
+pub mod resolver;
+pub mod storage_impl;
+
+pub use crate::core::FungibleTokenCore;
+pub use core_impl::{Balance, FungibleToken};
+pub use resolver::FungibleTokenResolver;
