+++ near-ft-standard/src/metadata.rs
