+++ near-ft-standard/src/receiver.rs
