+++ near-ft-standard/src/resolver.rs
